create method for verify user and email at create account

+++ model/delete_account.py
@@ -1,4 +1,6 @@
 from model.connection import *
+from os import *
+import sys
 class Delete():
     """class for user delete this account"""
     def __init__(self, pseudo):
@@ -7,7 +9,15 @@ class Delete():
 
     def del_account(self):
         """"method for delte user account after connect to bdd"""
-        self.user_choice.initialize_connection()
-        self.user_choice.cursor.execute("DELETE FROM users WHERE pseudo = %s;", (self.pseudo,))
-        self.user_choice.connection.commit()
-        self.user_choice.close_connection()
+        confirm = input("delete account y/n:")
+        if confirm == "y":
+            self.user_choice.initialize_connection()
+            self.user_choice.cursor.execute("DELETE FROM users WHERE pseudo = %s;", (self.pseudo,))
+            self.user_choice.connection.commit()
+            self.user_choice.close_connection()
+            print("account deleted good bye")
+            #system('clear')
+            sys.exit()
+        else:
+            system('clear')
+            return
